--- a/src/algorithms/div/knuth.rs
+++ b/src/algorithms/div/knuth.rs
@@ -1,9 +1,11 @@
 //! Knuth division
 
 use super::{reciprocal::reciprocal_2, small::div_3x2, DoubleWord};
-use crate::algorithms::{add::adc_n, mul::submul_nx1};
-use core::{ u64};
-use crate::utils::unlikely;
+use crate::{
+    algorithms::{add::adc_n, mul::submul_nx1},
+    utils::unlikely,
+};
+use core::u64;
 
 /// In-place Knuth long division
 #[allow(clippy::many_single_char_names)]