```
div_rem/64/64           time:   [2.1268 ns 2.1539 ns 2.1845 ns]
                        change: [-72.862% -71.771% -70.465%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe

div_rem/64/32           time:   [2.1675 ns 2.1961 ns 2.2279 ns]
                        change: [-52.439% -48.445% -42.205%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe

div_rem/64/64 #2        time:   [2.1259 ns 2.1543 ns 2.1857 ns]
                        change: [-73.510% -70.290% -66.428%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

div_rem/128/64          time:   [12.650 ns 12.717 ns 12.786 ns]
                        change: [-50.428% -48.756% -45.771%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  6 (6.00%) high mild
  2 (2.00%) high severe

div_rem/128/64 #2       time:   [12.444 ns 12.572 ns 12.743 ns]
                        change: [-50.022% -48.978% -47.450%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  6 (6.00%) high mild
  5 (5.00%) high severe

div_rem/128/128         time:   [29.542 ns 29.625 ns 29.718 ns]
                        change: [-40.395% -39.007% -37.104%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

div_rem/192/64          time:   [18.469 ns 18.644 ns 18.823 ns]
                        change: [-62.757% -61.367% -59.395%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

div_rem/192/96          time:   [46.363 ns 46.588 ns 46.833 ns]
                        change: [-49.458% -48.498% -47.340%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

div_rem/192/192         time:   [45.758 ns 45.943 ns 46.139 ns]
                        change: [-17.632% -15.498% -12.510%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

div_rem/256/64          time:   [22.147 ns 22.417 ns 22.694 ns]
                        change: [-67.825% -66.258% -63.647%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe

div_rem/256/128         time:   [42.537 ns 42.767 ns 43.013 ns]
                        change: [-62.118% -61.088% -59.949%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  8 (8.00%) high mild
  2 (2.00%) high severe

div_rem/256/256         time:   [48.165 ns 48.314 ns 48.462 ns]
                        change: [-20.454% -18.078% -15.220%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

div_rem/384/64          time:   [34.251 ns 34.668 ns 35.040 ns]
                        change: [-71.742% -70.733% -69.349%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

div_rem/384/192         time:   [83.828 ns 84.101 ns 84.384 ns]
                        change: [-46.296% -45.122% -43.775%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

div_rem/384/384         time:   [52.547 ns 52.797 ns 53.072 ns]
                        change: [-27.061% -24.922% -23.323%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

div_rem/512/64          time:   [48.732 ns 49.171 ns 49.544 ns]
                        change: [-71.472% -70.558% -69.155%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

div_rem/512/256         time:   [108.77 ns 109.13 ns 109.49 ns]
                        change: [-45.894% -45.140% -44.381%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

div_rem/512/512         time:   [56.192 ns 56.485 ns 56.785 ns]
                        change: [-26.234% -23.821% -19.723%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

div_rem/4096/64         time:   [547.19 ns 550.01 ns 552.59 ns]
                        change: [-65.041% -64.151% -62.870%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

div_rem/4096/2048       time:   [1.9356 µs 1.9418 µs 1.9484 µs]
                        change: [-25.095% -24.704% -24.265%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe

div_rem/4096/4096       time:   [272.14 ns 275.96 ns 279.29 ns]
                        change: [-37.468% -34.359% -31.871%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe
```